m4_proj_dbt_taxi_rides_ny/pyproject.toml
[project]
name = "m4-proj-dbt-taxi-rides-ny"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-duckdb>=1.10.0",
    "tqdm>=4.67.1",
]
